+++ packages/desktop-client/src/components/manager/BudgetList.tsx
@@ -16,13 +16,7 @@ import { Tooltip } from '@actual-app/components/tooltip';
 import { View } from '@actual-app/components/view';
 
 import {
-  closeAndDownloadBudget,
-  closeAndLoadBudget,
-  createBudget,
-  downloadBudget,
   getUserData,
-  loadAllFiles,
-  loadBudget,
   pushModal,
 } from 'loot-core/client/actions';
 import {
@@ -55,6 +49,7 @@ import { theme } from '../../style';
 import { tokens } from '../../tokens';
 import { useResponsive } from '../responsive/ResponsiveProvider';
 import { useMultiuserEnabled } from '../ServerContext';
+import { closeAndDownloadBudget, closeAndLoadBudget, createBudget, downloadBudget, loadAllFiles, loadBudget } from 'loot-core/client/budgets/budgetsSlice';
 
 function getFileDescription(file: File, t: (key: string) => string) {
   if (file.state === 'unknown') {
@@ -561,14 +556,14 @@ export function BudgetList({ showHeader = true, quickSwitchMode = false }) {
 
     if (!id) {
       if (isRemoteFile) {
-        await dispatch(downloadBudget(file.cloudFileId));
+        await dispatch(downloadBudget({ cloudFileId: file.cloudFileId }));
       } else {
-        await dispatch(loadBudget(file.id));
+        await dispatch(loadBudget({ id: file.id }));
       }
     } else if (!isRemoteFile && file.id !== id) {
-      await dispatch(closeAndLoadBudget(file.id));
+      await dispatch(closeAndLoadBudget({ fileId: file.id }));
     } else if (isRemoteFile) {
-      await dispatch(closeAndDownloadBudget(file.cloudFileId));
+      await dispatch(closeAndDownloadBudget({ cloudFileId: file.cloudFileId }));
     }
   };
