+++ packages/desktop-client/src/components/manager/ConfigServer.tsx
@@ -6,7 +6,8 @@ import { Button, ButtonWithLoading } from '@actual-app/components/button';
 import { Text } from '@actual-app/components/text';
 import { View } from '@actual-app/components/view';
 
-import { createBudget, loggedIn, signOut } from 'loot-core/client/actions';
+import { loggedIn, signOut } from 'loot-core/client/actions';
+import { createBudget } from 'loot-core/client/budgets/budgetsSlice';
 import {
   isNonProductionEnvironment,
   isElectron,
