+++ packages/desktop-client/src/components/manager/WelcomeScreen.tsx
@@ -7,7 +7,8 @@ import { styles } from '@actual-app/components/styles';
 import { Text } from '@actual-app/components/text';
 import { View } from '@actual-app/components/view';
 
-import { createBudget, pushModal } from 'loot-core/client/actions';
+import { pushModal } from 'loot-core/client/actions';
+import { createBudget } from 'loot-core/client/budgets/budgetsSlice';
 
 import { useDispatch } from '../../redux';
 import { theme } from '../../style';
@@ -97,7 +98,7 @@ export function WelcomeScreen() {
           <Button
             variant="primary"
             autoFocus
-            onPress={() => dispatch(createBudget())}
+            onPress={() => dispatch(createBudget({}))}
           >
             {t('Start fresh')}
           </Button>
