+++ packages/desktop-client/src/components/modals/CreateEncryptionKeyModal.tsx
@@ -11,8 +11,9 @@ import { Text } from '@actual-app/components/text';
 import { View } from '@actual-app/components/view';
 import { css } from '@emotion/css';
 
-import { loadAllFiles, loadGlobalPrefs } from 'loot-core/client/actions';
+import { loadGlobalPrefs } from 'loot-core/client/actions';
 import { sync } from 'loot-core/client/app/appSlice';
+import { loadAllFiles } from 'loot-core/client/budgets/budgetsSlice';
 import { send } from 'loot-core/platform/client/fetch';
 import { getCreateKeyError } from 'loot-core/shared/errors';
