+++ packages/desktop-client/src/components/modals/OutOfSyncMigrationsModal.tsx
@@ -6,7 +6,7 @@ import { Paragraph } from '@actual-app/components/paragraph';
 import { Text } from '@actual-app/components/text';
 import { View } from '@actual-app/components/view';
 
-import { closeBudget } from 'loot-core/client/actions';
+import { closeBudget } from 'loot-core/client/budgets/budgetsSlice';
 
 import { useDispatch } from '../../redux';
 import { Link } from '../common/Link';
