+++ packages/desktop-client/src/components/modals/PasswordEnableModal.tsx
@@ -6,7 +6,8 @@ import { Label } from '@actual-app/components/label';
 import { styles } from '@actual-app/components/styles';
 import { View } from '@actual-app/components/view';
 
-import { closeBudget, popModal } from 'loot-core/client/actions';
+import { popModal } from 'loot-core/client/actions';
+import { closeBudget } from 'loot-core/client/budgets/budgetsSlice';
 import { send } from 'loot-core/platform/client/fetch';
 import * as asyncStorage from 'loot-core/platform/server/asyncStorage';
