+++ packages/desktop-client/src/components/modals/manager/DeleteFileModal.tsx
@@ -5,7 +5,7 @@ import { ButtonWithLoading } from '@actual-app/components/button';
 import { Text } from '@actual-app/components/text';
 import { View } from '@actual-app/components/view';
 
-import { deleteBudget } from 'loot-core/client/actions';
+import { deleteBudget } from 'loot-core/client/budgets/budgetsSlice';
 import { type File } from 'loot-core/types/file';
 
 import { useDispatch } from '../../../redux';
@@ -71,10 +71,10 @@ export function DeleteFileModal({ file }: DeleteFileProps) {
                   onPress={async () => {
                     setLoadingState('cloud');
                     await dispatch(
-                      deleteBudget(
-                        'id' in file ? file.id : undefined,
-                        file.cloudFileId,
-                      ),
+                      deleteBudget({
+                        id: 'id' in file ? file.id : undefined,
+                        cloudFileId: file.cloudFileId,
+                      }),
                     );
                     setLoadingState(null);
 
@@ -137,7 +137,7 @@ export function DeleteFileModal({ file }: DeleteFileProps) {
                   }}
                   onPress={async () => {
                     setLoadingState('local');
-                    await dispatch(deleteBudget(file.id));
+                    await dispatch(deleteBudget({ id: file.id }));
                     setLoadingState(null);
 
                     close();
