+++ packages/desktop-client/src/components/settings/index.tsx
@@ -6,7 +6,8 @@ import { Text } from '@actual-app/components/text';
 import { View } from '@actual-app/components/view';
 import { css } from '@emotion/css';
 
-import { closeBudget, loadPrefs } from 'loot-core/client/actions';
+import { loadPrefs } from 'loot-core/client/actions';
+import { closeBudget } from 'loot-core/client/budgets/budgetsSlice';
 import { listen } from 'loot-core/platform/client/fetch';
 import { isElectron } from 'loot-core/shared/environment';
