+++ packages/desktop-client/src/components/sidebar/BudgetName.tsx
@@ -8,7 +8,7 @@ import { Popover } from '@actual-app/components/popover';
 import { Text } from '@actual-app/components/text';
 import { View } from '@actual-app/components/view';
 
-import { closeBudget } from 'loot-core/client/actions';
+import { closeBudget } from 'loot-core/client/budgets/budgetsSlice';
 import * as Platform from 'loot-core/client/platform';
 
 import { useContextMenu } from '../../hooks/useContextMenu';
