+++ packages/desktop-client/src/index.tsx
@@ -16,6 +16,7 @@ import { createRoot } from 'react-dom/client';
 import * as accountsSlice from 'loot-core/client/accounts/accountsSlice';
 import * as actions from 'loot-core/client/actions';
 import * as appSlice from 'loot-core/client/app/appSlice';
+import * as budgetsSlice from 'loot-core/client/budgets/budgetsSlice';
 import * as queriesSlice from 'loot-core/client/queries/queriesSlice';
 import { runQuery } from 'loot-core/client/query-helpers';
 import { store } from 'loot-core/client/store';
@@ -36,6 +37,7 @@ const boundActions = bindActionCreators(
     ...actions,
     ...accountsSlice.actions,
     ...appSlice.actions,
+    ...budgetsSlice.actions,
     ...queriesSlice.actions,
   },
   store.dispatch,
