+++ packages/loot-core/src/client/actions/backups.ts
@@ -1,8 +1,7 @@
 // @ts-strict-ignore
 import { send } from '../../platform/client/fetch';
+import { closeBudget, loadBudget } from '../budgets/budgetsSlice';
 import { type AppDispatch, type GetRootState } from '../store';
-
-import { closeBudget, loadBudget } from './budgets';
 
 // Take in the budget id so that backups can be loaded when a budget
 // isn't opened
@@ -14,7 +13,7 @@ export function loadBackup(budgetId, backupId) {
     }
 
     await send('backup-load', { id: budgetId, backupId });
-    await dispatch(loadBudget(budgetId));
+    await dispatch(loadBudget({ id: budgetId }));
   };
 }
