+++ packages/loot-core/src/client/actions/index.ts
@@ -1,6 +1,5 @@
 export * from './modals';
 export * from './notifications';
 export * from './prefs';
-export * from './budgets';
 export * from './backups';
 export * from './user';
