+++ packages/loot-core/src/client/actions/user.ts
@@ -1,8 +1,8 @@
 import { send } from '../../platform/client/fetch';
+import { closeBudget, loadAllFiles } from '../budgets/budgetsSlice';
 import * as constants from '../constants';
 import { type AppDispatch } from '../store';
 
-import { loadAllFiles, closeBudget } from './budgets';
 import { loadGlobalPrefs } from './prefs';
 
 export function getUserData() {
